Change the WifiConnectionSetupFragment to be like the BluetoothConnectionSetupFragment. You Choose your opponent from a list. Both players need to hold the item of the desired opponent. Determining who will connect to the other is base of comparing their nsd service names, which are unique in each private network by design of the nsd register service protocol. Change the DeviceAdapter to be more generic and hold list of Objects instead of using different adapters for wifi and bluetooth. Remove the old WifiConnectionSetupFragment.

+++ app/src/main/java/com/android/ronoam/taps/GameLogic/TypeLogic.java
@@ -44,7 +44,7 @@ public class TypeLogic implements TextWatcher {
 
     public float getOpponentResults(){
         int timeSeconds = (int)FinalVariables.TYPE_GAME_TIME / 1000;
-        return timeSeconds < 60 ? (60/timeSeconds) * opponentCounter : (timeSeconds/60) * opponentCounter;
+        return timeSeconds < 60 ? (60f/timeSeconds) * (float)opponentCounter : (timeSeconds/60f) * opponentCounter;
     }
 
     /*public String getNextWord(){
